Adding initial delay
When a stepper motor is turned on it is possible the rotor is not lined up with the magnetic field. In these first few moments of initializing the motor will line up again. This process takes some time; if you start controlling the motor within this time you can get unexpected behaviour. The problem is that the unallingment is everytime different therefore it can happen that one time a profile works and another time it wont.
I had a lot of problems with this effect. I edited this library to have a starting delay after initializing and got my project to work.